### 2021-11-12 #3909 @kyleniemeyer
Hi @bramvandijk88, as we are getting started in this pre-review issue, could you fix those missing DOIs in the references?

@Bisaloo could you edit this submission?
