### 2021-11-14 #3909 @bramvandijk88
> Hi @bramvandijk88, as we are getting started in this pre-review issue, could you fix those missing DOIs in the references?
>
> @Bisaloo could you edit this submission?

Absolutely, I'll get on it.

On a seperate note, it's hard to find reviewers for this I can imagine. Not a lot of javascript programmers are in your list. In any case, these usernames seemed somewhat appropriate for my submission:

emilydolson
kavir1698
jamesdamillington
TimKam
