### 2021-11-15 45515d8
Fixed rare issue with RNG in TE example
