+++ lib/mersenne_browser.js
@@ -179,10 +179,17 @@ function MersenneTwister(seed) {
   }
   
   /* generates a random number on [0,1)-real-interval */
+  MersenneTwister.prototype.genrand_real2 = function() {
+    return this.genrand_int32()*(1.0/4294967296.0); 
+    /* divided by 2^32 */
+  }
+
+  /* generates a random number on [0,1)-real-interval */
   MersenneTwister.prototype.random = function() {
     return this.genrand_int32()*(1.0/4294967296.0); 
     /* divided by 2^32 */
   }
+  
    
   /* generates a random number on (0,1)-real-interval */
   MersenneTwister.prototype.genrand_real3 = function() {
