### 2021-11-15 8eab7e2
Updated the comment-string above make_bundle to clarify what it isn't meant for
