+++ README.md
@@ -34,7 +34,6 @@ Complex systems like microbial communities, the human immune system, and the Ear
 
 The bundle was made with rollup:
 > rollup src/model.js -o dist/cacatoo.js -f cjs  -w
-(also see make_bundle.sh)
 
 Documentation was compiled with jsdoc (npm install jsdoc -g)
 > ./node_modules/.bin/jsdoc dist/cacatoo.js -d docs/jsdocs
