+++ paper/paper.bib
@@ -1,10 +1,12 @@
 
+
 @article{starruss2014morpheus,
   title={Morpheus: a user-friendly modeling environment for multiscale and multicellular systems biology},
   author={Starru{\ss}, J{\"o}rn and De Back, Walter and Brusch, Lutz and Deutsch, Andreas},
   journal={Bioinformatics},
   volume={30},
   number={9},
+  doi={10.1093/bioinformatics/btt772},
   pages={1331--1332},
   year={2014},
   publisher={Oxford University Press}
@@ -15,6 +17,7 @@
   author={Wortel, Inge MN and Textor, Johannes},
   journal={Elife},
   volume={10},
+  doi={10.7554/eLife.61288},
   pages={e61288},
   year={2021},
   publisher={eLife Sciences Publications Limited}
@@ -31,6 +34,7 @@
   title={Transposable elements drive the evolution of genome streamlining},
   author={van Dijk, Bram and Bertels, Frederic and Stolk, Lianne and Takeuchi, Nobuto and Rainey, Paul B},
   journal={bioRxiv},
+  doi={10.1101/2021.05.29.446280},
   year={2021},
   publisher={Cold Spring Harbor Laboratory}
 }
@@ -39,6 +43,7 @@
   title={Resolving structure and function of metaorganisms through a holistic framework combining reductionist and integrative approaches},
   author={Jaspers, Cornelia and Fraune, Sebastian and Arnold, A Elizabeth and Miller, David J and Bosch, Thomas CG and Voolstra, Christian R and others},
   journal={Zoology},
+  doi={10.1016/j.zool.2019.02.007},
   volume={133},
   pages={81--87},
   year={2019},
@@ -50,6 +55,7 @@
   author={Rainey, Paul B and Quistad, Steven D},
   journal={Philosophical Transactions of the Royal Society B},
   volume={375},
+  doi={10.1098/rstb.2019.0248},
   number={1798},
   pages={20190248},
   year={2020},
@@ -62,6 +68,7 @@
   journal={bioRxiv},
   pages={053405},
   year={2016},
+  doi={10.1101/053405},
   publisher={Cold Spring Harbor Laboratory}
 }
 @article{fullmer2015pan,
@@ -69,6 +76,7 @@
   author={Fullmer, Matthew S and Soucy, Shannon Margaret and Gogarten, Johann Peter},
   journal={Frontiers in microbiology},
   volume={6},
+  doi={10.3389/fmicb.2015.00728},
   pages={728},
   year={2015},
   publisher={Frontiers}
