### 2021-11-15 #3909 @Bisaloo
:wave: :wave: @kyleniemeyer @bramvandijk88, yes I can edit this issue.

Before we can get properly started, I notice that whedon severely overestimates the number of code lines. Whedon uses GitHub's linguist to detect the relevant files & languages. Could you edit your `.gitattributes` with the relevant `linguist-` attributes ([examples](https://stackoverflow.com/a/43142088)) so that auto-generated files, external libraries, etc. are excluded from the line count please?

Should the `node_modules` folder be committed anyways? The `package-lock.json` should be enough.
