### 2021-11-16 #3909 @bramvandijk88
> 👋 👋 @kyleniemeyer @bramvandijk88, yes I can edit this issue.

Thanks for editing.

> Before we can get properly started, I notice that whedon severely overestimates the number of code lines. Whedon uses GitHub's linguist to detect the relevant files & languages. Could you edit your `.gitattributes` with the relevant `linguist-` attributes ([examples](https://stackoverflow.com/a/43142088)) so that auto-generated files, external libraries, etc. are excluded from the line count please?
>
> Should the `node_modules` folder be committed anyways? The `package-lock.json` should be enough, shouldn't it?

Yeah, good point. I didn't even notice I had added the node modules, so I'll get on it.

> I couldn't find good documentation on how to use linguist so if this is not clear, please let me know and I'll try to submit a PR myself.

I'll let you know when I've updated it, and let's see.
