### 2021-11-16 #3909 @bramvandijk88
@Bisaloo I have

i) removed node modules from the repository and added it to .gitignore
ii) added a .gitattributes file that excludes certain files from the documentation that aren't part of the core code, or are duplicated files for the github.io page.

Let me know if this resolves the line count issues.
