### 2021-11-16 #3909 @Bisaloo
Thanks for your quick answer! I believe a couple of other changes are necessary:

- `dist/` should be `.gitignore`d since users will generate it from the contents of `src/`
- `lib/` contents are third party dependencies, right? If so, they should be marked as `linguist-vendored` in `.gitignore`

Please let me know if I misunderstand something or if you disagree.

I believe we should be good to go after this :+1:.
