### 2021-11-16 #3909 @Bisaloo
Regarding reviewers, we are not limited the google spreadsheet. The list contains people who agreed to review for JOSS but if you know potential good reviewers who are not in the list, feel free to suggest them as well.
