### 2021-11-16 #3909 @bramvandijk88
> Thanks for your quick answer! I believe a couple of other changes are necessary:
>
> * `dist/` should be `.gitignore`d since users will generate it from the contents of `src/`

No, I don't actually want to do that. I'm uploading the compiled library to the repository so that users can immediately dive into the examples without having to install anything. The zero-install is one of the selling points of this package ;D

> * `lib/` contents are third party dependencies, right? If so, they should be marked as `linguist-vendored` in `.gitignore`

Yes, thanks, I've added it.

>
> Please let me know if I misunderstand something or if you disagree.
>
> I believe we should be good to go after this 👍.

Alright, good. The problem with finding reviewers is that this is a really niche library (javascript is not very common, and the type of models are also quite niche), so I don't know who to suggest except my close colleagues, which is clearly a conflict of interest...
