### 2021-11-16 #3909 @Bisaloo
Okay, fair enough. I had a look at the jsfiddle examples and it looks very nice.

I would still recommend you add `linguist-generated` to `dist/` so it doesn't show up in diff but it's up to you.

I'll start looking into reviewers.
