### 2021-11-16 #3909 @bramvandijk88
> I would still recommend you add `linguist-generated` to `dist/` so it doesn't show up in diff but it's up to you.

Actually, yes, I'll do that.
