### 2021-11-17 5d8400c
Updated white colour in graph.
