### 2021-11-18 #3909 @Bisaloo
Awesome, thank for your quick answer :100:. Once I find the 2nd reviewer, I'll open the review thread and explain how it works.

If you are already curious, all the information is summarized in our documentation at https://joss.readthedocs.io/en/latest/.
