### 2021-11-18 #3909 @TimKam
I still need to review another JOSS paper (2nd round), but in roughly a week, I should be available. I had a quick look and I'd look forward to reviewing this project.
