### 2021-11-24 #3948 @Bisaloo
@emilydolson, yes, you should be able to edit the checklist once you have accepted the invitation. I'll re-invite you :relaxed:
