### 2022-01-12 #3948 @emilydolson
Yes! Apologies for being slow - my university switched teaching modalities and I've been scrambling to adjust my course.
