### 2022-01-12 bramvandijk88/cacatoo#14 @emilydolson
JOSS Review: Improvements to testing architecture
According to the JOSS guidelines, ideally your code should have automated tests that run on new commits (e.g. via Github Actions or another continuous integration platform). At a minimum it should have documentation for how to assess that the code is working as intended.

Currently it looks like Cacatoo has a unittests directory with an html file and a javascript file. That's a good start, but it's not quite sufficient. At a minimum, the expected behavior for those files should be documented somewhere so that a user can confirm that what they observe matches that. A better approach would be to include some assertions in your javascript so that it will throw an error if it doesn't behave as its supposed to. A Javascript testing framework might make this easier (I happen to like [mocha](https://mochajs.org/), but there are many options).

(This issue is part of my review for JOSS: https://github.com/openjournals/joss-reviews/issues/3948)
