### 2022-01-12 bramvandijk88/cacatoo#15 @emilydolson
JOSS Review: Contribution/community Guidelines
According to the JOSS Guidelines:

> There should be clear guidelines for third-parties wishing to:
>
>     Contribute to the software
>     Report issues or problems with the software
>     Seek support
>

I'm currently not seeing those. The standard way to create them is to add a file to the top-level of your repo called CONTRIBUTING.md. Github has some nice templates for this that you can access by going to https://github.com/bramvandijk88/cacatoo/community.

(this issue is part of this JOSS review: https://github.com/openjournals/joss-reviews/issues/3948)
