### 2022-01-13 bramvandijk88/cacatoo#14 @bramvandijk88
Dear Emily Dolson,

I have thought about other testing possibilities (like Mocha), but decided it wasn't necessary (or useful even). I included a quasi "unit test" in the repository to make sure all the main functions work as intended, but other than that most of the programming comes from the user itself (iow Cacatoo not an independent piece of software, it's a framework!). In my experience with bioinformatic/computational tools, tests are most convenient to make sure everything is installed correctly (which isn't relevant for a JS bundle), or to test the expected input-output relations of functions (which my "unit test" actually does). So I think I've got the important stuff covered. If you disagree, let me know on what level you would implement more testing, and I can work on that.

Best,

Bram
