### 2022-01-13 bramvandijk88/cacatoo#15 @bramvandijk88
On the HTML page on which the Cacatoo models are displayed, there is a footer which states "Issues and suggestions can be reported here", which links to the Github issue page. That said, I agree some notes on this were missing in the repository, and I added a minimal CONTRIBUTING.MD file.

Thanks for that suggestion, that was totally missing! :)
