### 2022-01-14 bramvandijk88/cacatoo#14 @emilydolson
I totally agree that tests to make sure that things are installed correctly don't make sense for a javascript framework like Cacatoo, and that they key tests to have are about the input-output relations of functions. My concern is that the current unit test doesn't seem like it enforces very many input-output relations. Maybe I'm misunderstanding how it works, but it seems like line 99 and line 106 are the only places that it actually verifies that a particular behavior occurred. Obviously we can see what the other reviewer and the editor think, but my perspective is that it would be ideal to have a lot more points where you explicitly check for expected behavior. Currently it doesn't seem like the unit test will catch most bugs that don't directly produce errors.

Certainly when I personally think about adopting Cacatoo as a user (which I could imagine doing for some projects - it seems like a very nice tool!), the lack of thorough tests would give me pause. I currently contribute to the development of a different framework for web-based scientific software, and after seeing how many bugs our unit tests have caught, I have a hard time trusting any code (including my own!) without a solid testing suite.

My recommendation would be:

# Minimally
Have a unit test for each Cacatoo-provided function that the user is expected to call and which does not have an immediate visual effect. This would include all of the neighborhood retrieval functions, the grid events, the ODE solver, `this.synchronous`, `this.asynchronous`, etc. A side benefit is that this would create example of how to call all of those functions, which would also be helpful.

# Ideally
- Also have tests for functions that have visual effects (e.g. adding a graph). I think the cost-benefit ratio on these is lower (since they are harder to test and it will usually be obvious if they break, rather than resulting in insidious bugs that no one is aware of), but it's still nice to know that your testing framework will tell you if you break them.
- Set up continuous integration so those tests run whenever you make a commit
