### 2022-01-14 bramvandijk88/cacatoo#16 @emilydolson
JOSS Review: Paper comments
Here are my comments on the paper submitted to JOSS (https://github.com/openjournals/joss-reviews/issues/3948).

This is a really cool project and I very much enjoyed reading the paper! Great work! The only remotely substantive changes I'd suggest are:

- Adding some discussion of how Cacatoo differs from Artistoo (i.e. under what circumstances would you want to use Cacatoo instead?).
- Adding some discussion of assumptions that Cacatoo makes/more precision about what the intended use case is. For example, the fact that there is only one grid that is constantly being updated (rather than a "current" grid and a "next" grid) means that if you aren't careful to have `nextState` only affect the focal cell, you can get artifacts based on the order the cells are updated in. You can usually get around this issue by using asynchronous updates, but I'm not positive that you always can, so it would probably be good to call it out explicitly in either the paper or the documentation.

# Minor typographical issues

- There is an extra set of parentheses around the second citation on line 33
- On line 35, parentheses are missing around "Figure 1".
- Line 74, there is a space between "d3." and "js" that I don't think is supposed to be there
- Is there any way to add a citation for CASH? The link to the author's page could easily break in the future.
