### 2022-01-14 bramvandijk88/cacatoo#17 @emilydolson
JOSS Review: Minor suggested documentation improvements
Lastly, I've got a couple suggested additions to the documentation. These aren't necessary in my mind, but I think they would be helpful and increase adoption of this framework (they certainly would have been helpful to me as I was testing it out :smile:):

- The blank project template is super helpful. I'd add an explicit link to it somewhere in the documentation so it's easier to find (or if it's already there, somewhere more prominent).
- Some of the comments in empty_project.html are currently wrong (e.g. `// Set half (50%) of the Gridmodel's grid points to 1 (alive)`), which makes it harder to figure out how it's working. I'd recommend fixing them.
- This is a bigger project, but more extensive documentation of what all the functions available to the user are would be very helpful. For example, this would include a list of all the methods of the random number generator.
- More specific documentation about what exactly functions are expecting as arguments and providing as returned values. It took me a while to figure out what the `getNeighbor` methods return, or what the third argument to `setGridpoint` should look like.

This issue is part of my JOSS Review (https://github.com/openjournals/joss-reviews/issues/3948#)
