### 2022-01-14 #3948 @emilydolson
Okay, done with my initial review! Really cool project! @bramvandijk88 already added the contributing guidelines, so in my mind the only two issues currently blocking acceptance are:

- [Improved unit tests](https://github.com/bramvandijk88/cacatoo/issues/14)
- [A little more discussion of the state of the field](https://github.com/bramvandijk88/cacatoo/issues/16)
