### 2022-01-14 bramvandijk88/cacatoo#14 @bramvandijk88
Thanks for the clear suggestions, that helps.

I’ll set up a more extensive test and also see if I can integrate it so it is ran upon commit. (That last one I’m a bit more skeptical on because in my experience such a setup is more likely to catch errors in your tests than errors in your code, which I find more than a bit frustrating, but I’ll give it a good old-fashioned try)
