+++ unit_test/mocha_unit_test.js
@@ -1,0 +1,214 @@
+// *********************************************************************************************************************
+// This Mocha unit test uses Chai to test whether the non-visual components of Cacatoo work as intended
+// *********************************************************************************************************************
+
+const chai = require('chai');
+const expect = chai.expect
+
+Simulation = require('../dist/cacatoo.js')
+
+describe('Cacatoo simulation unit test', function() 
+{  
+   describe('Testing grid model functionality...', function() 
+   {
+     it('Generate a simulation object, make an model with the name "grid model"', function() 
+     {          
+       const sim = new Simulation();       
+       const model = sim.makeGridmodel("gridmodel")
+       expect(sim).to.have.property('gridmodel')
+       expect(sim.gridmodel).to.be.an('object').that.respondsTo('nextState')
+       expect(sim.gridmodel.scale).to.be.above(0)
+       expect(sim.gridmodel.nc).to.be.above(0)
+       expect(sim.gridmodel.nr).to.be.above(0)       
+       expect(sim.gridmodel.graph_update).to.be.above(0)
+       expect(sim.gridmodel.graph_interval).to.be.above(0)
+       expect(sim.gridmodel.margolus_phase).to.be.above(-1)
+       expect(sim.gridmodel.grid.length).to.be.equal(sim.ncol)
+       expect(sim.gridmodel.grid[0].length).to.be.equal(sim.nrow)
+       expect(sim.gridmodel.grid[0][0]).to.be.an('object')
+       expect(() => sim.gridmodel.update()).to.throw();
+       sim.gridmodel.update = function() {}
+       sim.gridmodel.update()
+     });
+     it('Calling core functions for grid functionality', function() 
+     {          
+       const sim = new Simulation();       
+       sim.makeGridmodel("gridmodel")
+       sim.gridmodel.MargolusDiffusion()     
+       sim.gridmodel.perfectMix()       
+       expect(() => sim.gridmodel.nextState()).to.throw();
+       sim.gridmodel.nextState = function (i, j) { this.value = 1 }
+       sim.gridmodel.nextState()
+       sim.gridmodel.synchronous()
+       sim.gridmodel.asynchronous()
+     });
+   });
+
+   describe('Populating grid model...', function() 
+   {
+     const sim = new Simulation();       
+     sim.makeGridmodel("populate")
+     let individuals = [{alive:true, species:1, name: 'A'},{alive:true, species:2, name: 'B'}]
+     it('Testing initialGrid', function() 
+     {                
+       sim.initialGrid(sim.populate, 'species', 1, 0.33, 2, 0.33, 3, 0.33)
+       expect(sim.populate.grid[0][0]).to.have.property('species')       
+       sim.populate.clearGrid()
+       expect(sim.populate.grid[0][0] === sim.populate.grid[1][1]).to.equal(false)
+       expect(sim.populate.grid[0][0]).to.not.have.property('species') 
+      });
+
+      it('Testing initialSpot', function() 
+      {
+       sim.initialSpot(sim.populate, 'species', 1, 100, 0, 0)
+       expect(sim.populate.grid[0][0]).to.have.property('species')       
+       sim.populate.clearGrid()
+      });
+
+      it('Testing populateGrid', function() 
+      {       
+       sim.populateGrid(sim.populate, individuals, [0.5,0.5])
+       expect(sim.populate.grid[0][0]).to.have.property('species')      
+       expect(sim.populate.grid[0][0]).to.have.property('alive')
+       expect(sim.populate.grid[0][0]).to.have.property('name')
+       sim.populate.clearGrid()
+      });
+
[…]
