+++ test/bundle_test/bundle_test.js
@@ -1,149 +1,0 @@
-
-// With this "bundle test", most of the functionalities in Cacatoo are called, and it is checked whether any unexpected behaviour occurs. 
-
-if (typeof window == "undefined") Simulation = require('../dist/cacatoo.js') // Loads the Simulation class for nodejs-mode
-
-let sim;
-
-var birth = 1.0                    // Mutualist species A giving help to reproduce B
-var death = 0.1                    // Mutualist species B giving help to reproduce A
-
-
-function cacatoo() {
-    let config = {                                                      
-        title: "Simulation test",
-        description: "Unit test (sort of...) that checks if implemented functions are working as intended",
-        maxtime: 1000,
-        ncol: 100,
-        nrow: 100,		                                           
-        wrap: [true, true],                                        
-        seed: 56,
-        fps: 60,           
-        fpsmeter: false,
-        sleep: 1,        
-        fastmode: false,
-        scale: 2,				                                   
-        graph_interval: 10,
-        graph_update: 50,
-        statecolours: { 'discrete_var': { 1: "gold", 2: [255,0,0], 3: "#3030ff"}, // Setup colours with the 3 available options
-                        'continuous_var': 'viridis'                              // Setup colour gradient
-                      },
-        num_colours: 100,
-        skipbg_state: true,
-        show_gridname: true,
-        printcursor: false,
-    }
-
-    sim = new Simulation(config)                                              
-
-    sim.makeGridmodel("model")                                                
-    sim.initialGrid(sim.model, 'discrete_var', 1, 0.33, 2, 0.33, 3, 0.33)
-    sim.initialGrid(sim.model, 'continuous_var', 100, 1.0)
-    sim.initialGrid(sim.model, 'continuous_var2', 100, 1.0)
-
-    sim.createDisplay("model", "discrete_var", "Discrete variable")                    
-    sim.spaceTimePlot("model", "Discrete variable", "Space-time (discr)", 50, sim.ncol)                                  
-    sim.createDisplay("model", "discrete_var", "(zoom in on grid)", 20, 20, 10)            
-
-    sim.createDisplay_continuous({model:"model", property:"continuous_var", label:"Some continuous variable", minval:0, maxval:100}) 
-    sim.spaceTimePlot("model", "Some continuous variable", "Space-time (continuous)", 50, sim.ncol)                        
-    sim.createDisplay_continuous({model:"model", property:"continuous_var2", label:"Some continuous variable 2", minval:0, maxval:100, num_colours: 20, fill: 'inferno'}) 
-    
-
-   
-    sim.model.nextState = function (i, j)                               
-    {
-        let state = this.grid[i][j].discrete_var;
-        neigh = this.randomMoore8(this, i, j);               
-        sum = this.countMoore8(this, i, j, 'discrete_var',2);         
-            
-        if (state == 0)                                                 
-        {
-            
-            //conc = this.countMoore8(this, i, j, 'species',3);   
-            ran = this.rng.random();                              
-
-            if (this.rng.random() < birth)                        
-                this.grid[i][j].discrete_var = neigh.discrete_var 
-           
-        }
-
-        if (state == 1) 
-            this.grid[i][j].continuous_var = 100
-        else 
-            this.grid[i][j].continuous_var -= sum            
-
-
-        this.grid[i][j].continuous_var2 -= sum
-        if(this.grid[i][j].continuous_var2 < 0)this.grid[i][j].continuous_var2 = 0
[…]
