+++ make_bundle.sh
@@ -1,71 +1,67 @@
 #!/bin/bash 
 
-# This bundle script is NOT meant for regular developers. It bundles all the source code together,
-# and make sure the paths to the documentations are updated. If you want your changed to be 
-# implemented in the documentation, issue a pull request on github and I (Bram) will do the rest.
-# sudo bash make_bundle.sh once; git add *; git commit -a -m "Message"; git push
+# This script bundles, tests, and prepares the documentation for Cacatoo
+# Only the end-developer (currently: Bram van Dijk) needs to run this when uploading a fully new distribution of Cacatoo.
+# If you have suggested changes to the documentation, issue a pull request on github.
+# ( rollout command: sudo bash make_bundle.sh once; git add *; git commit -a -m "Message"; git push )
 # 
 
 compile_cacatoo()
-{               
-            
-            rollup src/simulation.js -o dist/cacatoo.js -f cjs                              # Use rollup to bundle the package as a single commonJS file
-            sed -i '$ d' dist/cacatoo.js
-            echo "
-            try
-            {
-                module.exports = Simulation;
-            }
-            catch(err)
-            {
-                // do nothing
-            }" >> dist/cacatoo.js
+{                           
+    rollup src/simulation.js -o dist/cacatoo.js -f cjs                              # Use rollup to bundle the package as a single commonJS file
+    sed -i '$ d' dist/cacatoo.js
+    echo "
+    try
+    {
+        module.exports = Simulation;
+    }
+    catch(err)
+    {
+        // do nothing
+    }" >> dist/cacatoo.js
 
-            ./node_modules/mocha/bin/mocha unit_test/                                       # Run mocha unit test to ensure the bundle works as intended
-            
-            if [[ "$?" -gt 0 ]]; then
-                echo -e "Mocha: at least one of the unit tests failed\t[ERROR]";
-                echo -e "Fix above issues before commiting the new bundle.\n\n"
-                exit 1
-            fi
+    ./node_modules/mocha/bin/mocha unit_test/                                       # Run mocha unit test to ensure the bundle works as intended
+    
+    if [[ "$?" -gt 0 ]]; then
+        echo -e "Mocha: at least one of the unit tests failed\t[ERROR]";
+        echo -e "Fix above issues before commiting the new bundle.\n\n"
+        exit 1
+    fi
 
-            ./node_modules/.bin/jsdoc dist/cacatoo.js -d docs/jsdocs -R README.md           # Automatically recompile JSdocs
+    ./node_modules/.bin/jsdoc dist/cacatoo.js -d docs/jsdocs -R README.md           # Automatically recompile JSdocs
 
-            
-            cp examples/03_expert/legend.png docs/TEs_streamlining                          # Everything below = fixing documentation files
-            cp dist/cacatoo.js docs/scripts/cacatoo.js
-            cp style/cacatoo.css docs/styles/cacatoo.css
-            cp lib/all.js docs/scripts/all.js
-            cp lib/odex.js docs/scripts/odex.js
-            cp images/elephant_cacatoo_small.png docs/images/
-            cp images/elephant_cacatoo.png docs/images/
-            cp images/gh.png docs/images/
+    
+    cp examples/03_expert/legend.png docs/TEs_streamlining                          # Everything below = fixing documentation files
[…]
