### 2022-01-18 a364704
Added unit tests
