### 2022-01-18 bramvandijk88/cacatoo#14 @bramvandijk88
Alright, I've just pushed a new commit in which I use Mocha and Chai to test all **non-visual** functions of Cacatoo. This luckily didn't reveal any bugs (which is a good thing... I've spent months testing and debugging this library manually).

I've also updated the compilation script so these tests are performed every time I distribute a new version of the bundle. If I (or a user through a pull request) ever breaks something, I should automatically know. I do not want to burden users with mandatory testing though, although they are free to run the Mocha tests themselves if they want.

Finally, although I agree the cost/benefit ratio isn't very high to do the same with the visual components, the bundle_test (which was already present at previous commits) will be able to catch most of those too.
