### 2022-01-18 bramvandijk88/cacatoo#18 @Bisaloo
Generate docs automatically
I was trying to find the changes made as part of #14 but the git history is quite difficult to navigate at the moment. One reason for this is that actual changes are tangled with update of the docs build time within a single commit (e.g., https://github.com/bramvandijk88/cacatoo/commit/be3a41245487b0f892b9652578ec48c0382a1fa9).

Additionally, the process to generate docs is not documented at the moment. This makes it difficult for an external user to contribute or for a new maintainer to take over (if you ever get bored of this project or if you have other commitments that force you to let it go).

I believe a good option to:

- make the git history clearer
- make the doc building process apparent

would be to use a GitHub Action to automatically generate the docs. The JSDoc README suggests this action: https://github.com/andstor/jsdoc-action.

This issue is part of https://github.com/openjournals/joss-reviews/issues/3948.
