+++ paper/paper.bib
@@ -24,7 +24,7 @@
 }
 
 @book{zhu2013data,
-  title={Data visualization with D3. js cookbook},
+  title={Data visualization with D3.js cookbook},
   author={Zhu, Nick Qi},
   year={2013},
   publisher={Packt Publishing Ltd}
