+++ paper/paper.md
@@ -24,7 +24,7 @@ In the field of ecology and evolution, a treasure trove of data has revealed the
 
 # Statement of need
 
-Complex systems like microbial communities have many emergent properties which arise from the interactions between its individual components. As such, predicting exactly how these systems will behave and respond to various stimuli is difficult. Simulation offers a solution by allowing a modeller to simply put in what they deem important, and observe the outcome. As such, direct visual feedback is an important part of exploring the model. Not only is one more likely to detect programming mistakes this way, but it also aids in communicating the models to peers and the general public. Many current modelling frameworks are based in C (e.g. [Cash](https://tbb.bio.uu.nl/rdb/software.html) and [Morpheus](https://academic.oup.com/bioinformatics/article-abstract/30/9/1331/234757) [@starruss2014morpheus]), motivated by its unparalled speed. However, the learning curve for programming in C is steep, and even an experienced user may take days to track down a simple bug. Moreover, sharing your model with other users can be a pain, as installation is slightly different depending on each operating system. With Javascript tools like d3.js [@zhu2013data] and Artistoo ([@wortel2021artistoo]) paving the way, `Cacatoo` was developed to overcome these issues, making spatially structured models easy, fast, sharable, and customisable. The basic recipe of a Cacatoo simulation is simple \autoref{fig:recipe}, and the extensive documentation comes with plenty of examples to get started right away!<br><br>
+Complex systems like microbial communities have many emergent properties which arise from the interactions between its individual components. As such, predicting exactly how these systems will behave and respond to various stimuli is difficult. Simulation offers a solution by allowing a modeller to simply put in what they deem important, and observe the outcome. As such, direct visual feedback is an important part of exploring the model. Not only is one more likely to detect programming mistakes this way, but it also aids in communicating the models to peers and the general public. Many current modelling frameworks are based in C (e.g. [Cash](https://tbb.bio.uu.nl/rdb/software.html) and [Morpheus](https://academic.oup.com/bioinformatics/article-abstract/30/9/1331/234757) [@starruss2014morpheus]), motivated by its unparalled speed. However, the learning curve for programming in C is steep, and even an experienced user may take days to track down a simple bug. Moreover, sharing your model with other users can be a pain, as installation is slightly different depending on each operating system. With Javascript tools like d3.js [@zhu2013data] and Artistoo [@wortel2021artistoo] paving the way, `Cacatoo` was developed to overcome these issues, making spatially structured models easy, fast, sharable, and customisable. Other than Artistoo, which first and foremost specialises on tissue simulations, `Cacatoo` provides easy-to-use functions for individual-based modeling  (roulette wheel selection, random walks, neighbourhood retreival), reaction diffusion systems (ordinary differntial equations), and tools to build your own user interface (displays, graphs, sliders, interaction with mouse and keyboard, *etc.*). The basic recipe of a Cacatoo simulation is simple (\autoref{fig:recipe}), and the extensive documentation provides ample opportunity for students and advanced modellers to get started right away! <br><br>
 
 ![The basic recipe of a Cacatoo simulation contains three ingredients: 1) setup, 2) defining the rules, and 3) setting up the main simulation loop.\label{fig:recipe}](../images/cacatoo_recipe.png)
 
