### 2022-01-18 bramvandijk88/cacatoo#16 @bramvandijk88
I have just made some minor text changes to cover these issues. I have clarified what Cacatoo brings to the table compared to Artistoo in the paper:

> Other than Artistoo, which first and foremost specialises on tissue simulations, Cacatoo provides easy-to-use functions for individual-based modeling (roulette wheel selection, random walks, neighbourhood retreival), reaction diffusion systems (ordinary differntial equations), and tools to build your own user interface (displays, graphs, sliders, interaction with mouse and keyboard, etc.).

... and fixed the minor typographical issues you mentioned.

Next, you are absolutely on-point about the fact that I do not go into detail on what it entails to work with a grid-based models (e.g. synchronous vs asynchronous updating, as you explain), but I feel like it would require quite a few words to really go into that topic (it also feels out of place in the paper). Instead, I have elaborated on this point in the introducing tutorial (https://bramvandijk88.github.io/cacatoo/first_project.html):

> Note that synchronous updating necessitates that one does **not** modify the neighbouring grid points in the nextState function, as this would mean a grid point that was already updated can be modified again by one of its neighbouring grid points. If you want to modify neighbouring grid points in the nextState function, always make sure to update the grid asynchronously (using _this.asynchronous()_ instead of _this.synchronous()_), which updates all grid points in a random order.

I have also added an FAQ that may cover this particular type of artifact.

About the CASH-ciatation: there is, unfortunately, no official publication of CASH. The first version is from the 80s, and none of the updates since made it into an official methods paper.
