### 2022-01-18 bramvandijk88/cacatoo#18 @bramvandijk88
I am sorry the git history is a bit of a mess, this project was much more extensive than I initially anticipated. I have heard pretty scary stories about trying to clean up git histories, so I am not very enthusiastic about doing this.

In any case, the documentation process is already fully automated, but I prefer to use my own workflow (and a custom bash script) for this. The script make_bundle.sh bundles, tests, and documents all functions within Cacatoo. Every time I think it is time to roll out a new distribution (i.e. a one-file bundle), I run this script to ensure everything is up-to-date. The JSdocs that are generated with this script can be found on the Github website: https://bramvandijk88.github.io/cacatoo/

I feel like I may be a bit old-fashioned in some regards, but I am slightly allergic to integrating it fully with Github Actions. I am confident that, if I ever need to yield the project to another user, I will be able to explain my workflow to this person.
