### 2022-01-18 bramvandijk88/cacatoo#17 @bramvandijk88
I have fixed the empty_project.html file. I agree with you that I'm not sure it is super useful, but I decided to leave it there anyway for when users prefer to start with a nice empty file. It's not like the empty project is "in the way" or anything ;D

About the last two points: are you asking for anything beyond the already provided JSdocs? https://bramvandijk88.github.io/cacatoo/jsdocs/index.html
