### 2022-01-18 2ef4f0f
Minor fixes to documentation
