### 2022-01-21 bramvandijk88/cacatoo#17 @emilydolson
You might have misread my comment about the empty_project.html template - I think it *is* helpful! I found it really useful in trying out Cacatoo. :smile:. Those changes look good!

For the third point, it is entirely possible that everything I was looking for is in the JSdocs and I just failed to find it. Is there a list of random number generator methods in there? I looked for a while and couldn't find it.

For the last point, I definitely would have found it helpful to have more detail beyond what is currently there. For instance, I can see in the documentation that the third argument to `setGridpoint` is a `Gridpoint`, but because classes in Javascript are so flexible, it wasn't obvious to me what a Gridpoint object should look like. An example would have been super helpful here.
