### 2022-01-21 bramvandijk88/cacatoo#16 @emilydolson
These updates look good to me! I think the documentation is a fine place to discuss the synchronous/asynchronous point
