### 2022-01-21 bramvandijk88/cacatoo#14 @emilydolson
This looks good, and I'm glad to hear there were no bugs! Thanks for adding the tests!

I agree that the bundle_test is a good way to know that the visual components are working. If you included a screenshot somewhere of what it is supposed to look like, that would make it even more useful in that capacity to someone who is new to the project.
