### 2022-01-21 bramvandijk88/cacatoo#14 @Bisaloo
@bramvandijk88, could you still please mention in the `CONTRIBUTING.md` guide how an external contributor could run the tests? If I followed correctly, they would have to re-compile the library with `make_bundle.sh`, correct?
