### 2022-01-21 bramvandijk88/cacatoo#14 @bramvandijk88
> I agree that the bundle_test is a good way to know that the visual components are working. If you included a screenshot somewhere of what it is supposed to look like, that would make it even more useful in that capacity to someone who is new to the project.

That's a great idea actually, I've added a screenshot in the directory!
