### 2022-01-21 bramvandijk88/cacatoo#18 @Bisaloo
Okay, fair enough for the GitHub Action. This is ultimately your choice.

I do believe it is however necessary to explicitly document in the repo how you re-generate the docs (i.e., run `make_bundle.sh`). The `CONTRIBUTING.md` guide might be a good place for it.

Yes, I agree trying to re-write git history is generally dangerous and should only be done as a last resort. A simple "fix" for the future would be to try to always run `make_bundle.sh` in a separate commit (if possible).
