### 2022-01-21 d7c5a48
Added a screenshot to the bundle_test directory so users know what it supposed to look like
