### 2022-01-21 bramvandijk88/cacatoo#14 @bramvandijk88
> @bramvandijk88, could you still please mention in the `CONTRIBUTING.md` guide how an external contributor could run the tests? If I followed correctly, they would have to re-compile the library with `make_bundle.sh`, correct?

Actually, no. As mentioned in the top comments of make_bundle, it is only intended to be used by the end-stage developer (so me, currently). However, as requested, I've added the command used to run the mocha tests to `CONTRIBUTING.md` in the section "pull requests", so that users can report the output of the test when making a pull request.
