### 2022-01-21 ce4b282
Modified README.md and CONTRIBUTING.md to clarify usage of developer commands
