### 2022-01-21 bramvandijk88/cacatoo#18 @bramvandijk88
Yeah, I will try and be more careful with making smaller commits in the future. Some of them got lumped together because I changed too much at the same time, and I'm sorry if things got a little messy.

The README.md already mentioned how the documentation is generated, but I've also added it to CONTRIBUTING.MD just in case someone would miss it.
