### 2022-01-21 bramvandijk88/cacatoo#18 @Bisaloo
Thanks! This issue is resolved as far as I'm concerned.
