### 2022-01-21 3a09ac0
Made hamburger menu in DOCs scrollable for small screens.
