### 2022-01-21 bramvandijk88/cacatoo#17 @bramvandijk88
Actually, you're right that the external libraries used (the mersenne-twister RNG, the odex library, dygraphs) are not in the JSdocs. These were downloaded precisely in the format as they were available on other repositories, and some of them are minimised or simply were missing the docstrings to begin with. In most cases, I think that's fine, as nobody needs to know how dygraphs works under the hood (and they can look it up if necessary).

However, you are on point about the RNG functions being really useful to explain. I've added a page in the DOCs to explain the functions shipped with Mersenne-Twister.

Finally, about the contents of the "gridpoint" objects, it's actually not predefined at all. Gridpoint is pretty much an empty object, which only comes with a few functions for copying properties from template-objects. One of the reasons I decided to use javascript for this tool is to allow users to put "whatever" into their grid points. I have modified the docstrings to clarify that a typical user will use the output of getGridpoint to feed to the third argument of setGridpoint. Indeed, there's a lot of flexibility in Javascript, which makes documenting things like this a bit.... fuzzy. :/
