### 2022-01-23 bramvandijk88/cacatoo#19 @TimKam
JOSS Review, some improvements
Hi @bramvandijk88, this issue contains all feedback of my review (https://github.com/openjournals/joss-reviews/issues/3948).

### Installation & Installation Instructions
Typically, one would like to **install** the library by running ``npm install`` or similar.
Even if many in the 'target audience' may prefer working in the HTML files of a copy of the repository, it would be good to document how to work with Cacatoo as a properly managed software artifact that is installed using npm.

Nit-pick: ``npm audit`` finds a low severity vulnerability; of course, in this context the vulnerability is most likely irrelevant.

### Automated tests
To run the tests, one first needs to run ``npm install -g mocha`` or similar, and then ``npm install chai odex``.
Consider managing the dependencies properly in the ``package.json`` file. Also, considering running the tests on a CI like GitHub Actions upon every commit/before every merge.

### Community guidelines
Guidelines "for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" are missing.

### Nit-picks, paper

#### Summary
* "is however lagging behind" -> consider removing "however", the contrast is already established by "Despite"
* "limited understanding on how" -> "limited understanding of how"?
* "to make it easy to design," -> "to design," (consider, neutral langauge)
* "Because it requires no installation" is a strong statement, because as a user of the toolbox, one arguably wants to work in the node.js ecosystem.

#### Statement of need
* "are based in C" -> "are written in C" or "require knowledge of C"?
* "and the extensive documentation provides ample opportunity for students and advanced modellers to get started right away!" Consider using neutral language and avoiding exclamation marks.

#### Use Cases
* "The latter model was used in *our* recent work". In the first section, you use "I". Generally, it is good to be consistent, so replace "I" with first person plural ("we").
* "Matthew Fullmer is currently in the process of exploring" -> consider: "Fullmer reports on work in progress that explores..."
