### 2022-01-23 #3948 @TimKam
I have completed the review and have created [an issue in the source code repository](https://github.com/bramvandijk88/cacatoo/issues/19) that outlines some shortcomings/improvements, with respect to some issues around installation & instructions, automated tests, community guidelines, and the software paper (minor comments).
Conceptually/w.r.t. scholarly effort, I assess the paper as reasonable (from an engineering point of view, I do not have any expertise in ecology and evolution).
