### 2022-01-23 bramvandijk88/cacatoo#19 @bramvandijk88
Thanks for a the useful feedback. I'll get on it asap. One quick question though: the contribution stuff was also mentioned by the other reviewer. Are you sure I am still missing something?

Anyway, thanks for all the practical engineering tips. Apposed to you, I am mostly an expert in the biology, and NOT the engineering, so it's very much appreciated!
