### 2022-01-23 bramvandijk88/cacatoo#19 @TimKam
I suppose there should be a 'Contributing' section or sth. similar in your README or documentation. If you have one, point me to it. (Perhaps I simply failed to find it.)
