### 2022-01-23 bramvandijk88/cacatoo#19 @bramvandijk88
> I suppose there should be a 'Contributing' section or sth. similar in your README or documentation. If you have one, point me to it. (Perhaps I simply failed to find it.)

There is a CONTRIBUTING.md file. I suppose I could make a reference to that file also in the README.md file. Thanks for clarifying.
