+++ .github/workflows/github_actions.yml
@@ -1,0 +1,22 @@
+Run Mocha test
+
+name: "GitHub Actions Test"
+on:
+    pull_request:
+        branches:
+            - master
+
+
+jobs:
+    test:
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v1
+            - name: 'Install Node'
+              uses: actions/setup-node@v1
+            - name: Install mocha
+              run: npm install -g mocha
+            - name: Install dependencies
+              run: npm install 
+            - name: "Run Test"
+              run: mocha unit_test
