+++ .github/workflows/github_actions.yml
@@ -19,7 +19,7 @@ jobs:
             - name: Install chai
               run: npm install --save-dev chai
             - name: Install odex
-              run: npm install -g odex
+              run: npm install --save-dev odex
             - name: Install dependencies
               run: npm install 
             - name: "Run Test"
