+++ README.md
@@ -28,22 +28,7 @@ If you prefer to learn by example, I have made dozens of different models for yo
 
 Complex systems like microbial communities, the human immune system, and the Earth's climate, have many emergent properties which arise from the interactions between individual components. As such, predicting exactly how these systems will behave and respond to various stimuli is difficult. Simulation offers a solution by allowing a modeller to simply put in what they deem important, and observe the outcome. As such, direct visual feedback is an important part of the process of "getting to know" your model. Not only are you more likely to detect programming mistakes, but it also aids exploration of parameter space. Over the last decade, most of my models were implemented by using the CASH C-library, which would allow direct visual feedback by means of the X11 library (R.J. de Boer & A.D. Staritsk). However, as it is based in C, developing models with CASH requires a lot of programming experience, and even an experienced user like myself can sometimes take days to track down a simple bug. Moreover, sharing your model with other users can be a pain in the neck, as installation is slightly different depending on the operating system. The amazing toolbox [Artistoo](https://artistoo.net/) has illustrated how Javascript can be both beginner-friendly and versatile. Hence, I decided that Javascript was the way forward! <br><br>
 
-# Other
-
-## Notes for developers
-
-Contributions, suggestions, and issues are very welcome. See CONTRIBUTING.md for more details.
-
-The bundle was made with rollup:
-> rollup src/model.js -o dist/cacatoo.js -f cjs  -w
-
-Documentation was compiled with jsdoc (npm install jsdoc -g)
-> jsdoc dist/cacatoo.js -d docs/jsdocs
-
-Unit testing is done with Mocha (npm install mocha)
-> mocha unit_test/   
-
-## FAQ
+# FAQ
 
 **Q**: My model isn't running, what can I do to find out what's wrong?\
 **A**: Open the developer console (CTRL+SHIFT+I in Google Chrome), and see which line of your code causes the problem. If it is an error in Cacatoo, please submit an issue on the Github repository. <br> <br>
@@ -56,6 +41,57 @@ make sure only the focal cell is modified by the nextState function. If not, the
 **A**: Open the developer console (CTRL+SHIFT+I in Google Chrome), and go to the 'profile' tab. You can run the code for a bit and observe which functions are taking most time. As a general rule, try and avoid build-in functions of javascript (e.g. 'reduce' to get the sum of an array), because despite being easier to use, they are quite a bit slower than a C-style piece of code where you loop over the values yourself. After you understand the model, and you simply want to rapidly run many simulations, I recommend running the code from the command line with nodeJS (also see the "cheater" example).  <br> <br>
 **Q**: Why is this toolbox called Cacatoo?\
 **A**: Cacatoo is an acronym for CAsh-like Cellular Automaton TOOlbox. However, as Cacatoo developed it became much more than just a tool for cellular automata. The name stuck, however. <br><br>
+
+# Contributing to Cacatoo
+
+As the sole developer of Cacatoo, I am eager to get help, suggestions, or use cases from others. For organisational reasons, I encourage everyone to always submit an official Github issue, and be sure to include the following details:
+
+## Reporting bugs
+
+When describing a bug, make sure to:
+* Describe the unexpected behaviour
+* Describe the desired behaviour
+* Include a reproducable example
+
+## Suggesting additions
+
+When making suggestions, make sure to:
+* Describe the missing feature
+* Describe or sketch the expected output
+
+## Pull requests
+
+When you want to actively contribute to Cacatoo, you can suggest to become a collaborator on Github.
+Implemented new branches may be merged with the main branch if the changes are expected to be useful to other. 
+
[…]
