### 2022-01-23 317da23
Delete CONTRIBUTING.md
