+++ paper/paper.bib
@@ -1,4 +1,13 @@
-
+@misc{hall2022introduction,
+  title={Introduction: the secret lives of microbial mobile genetic elements},
+  author={Hall, James PJ and Harrison, Ellie and Baltrus, David A},
+  journal={Philosophical Transactions of the Royal Society B},
+  volume={377},
+  number={1842},
+  pages={20200460},
+  year={2022},
+  publisher={The Royal Society}
+}
 
 @article{starruss2014morpheus,
   title={Morpheus: a user-friendly modeling environment for multiscale and multicellular systems biology},
