### 2022-01-23 723da69
Published NPM package of version 1.0.0
