### 2022-01-23 bramvandijk88/cacatoo#19 @bramvandijk88
Alright, you can now confirm that I have:

- [x] Published Cacatoo as an npm package (name: cacatoo). I have tested this uploaded package with a few of the provided examples, and everything seems to work as expected.
- [x] Set up a Github Action for unit testing (mocha) upon pull requests. For working on the main branch I prefer to run the tests in my own time (as discussed at an earlier stage of reviews with @bisaloo)
- [x] Considered your paper remarks and made fixes where I agreed. I decided to retain the exclamation mark and other somewhat informal styles, but I have accepted all your other changes as an improvement.
- [x] Merged CONTRIBUTE.md and README.md into a single file (README.md) so it's harder to overlook (i.e. it's on the main page now)

Once again thank you for your input, I definitely overlooked some things. I am happy with how polished the package has now become! :)

PS I noticed that Github does not automatically index the package as "published". Is there anything I can do to fix that?
