### 2022-01-23 0043652
Updated PNG for bundle_test
