### 2022-01-23 bramvandijk88/cacatoo#19 @TimKam
For some reason (probably: lack of permissions), I cannot tick the boxes in your comment. Anyway, I think the changes are okay. Regarding GitHub's 'published' information, you need to click the 'Create a new release' link and then tag the commit that reflects the state of the release and enter a description of the release. A minor change you can make is adding the CI status badge to the beginning of the README. You can also try to fix all security vulnerabilities in your npm release, but it's debatable whether what npm/GitHub tags as a 'vulnerability' is relevant in your context. What would be good in the long run (consider opening a new issue for this) is to create an example that demonstrates how to get started with cacatoo in a proper JavaScript project where it is installed via npm (included as a dependency). This could be one of the many examples you have already, merely with a slightly different setup.
