+++ dist/cacatoo.js
@@ -2586,12 +2586,4 @@ function get2DFromCanvas(canvas) {
     catch(err)
     {
         // do nothing
-
-    try
-    {
-        module.exports = Simulation;
-    }
-    catch(err)
-    {
-        // do nothing
     }
