### 2022-01-23 bramvandijk88/cacatoo#19 @bramvandijk88
> What would be good in the long run (consider opening a new issue for this) is to create an example that demonstrates how to get started with cacatoo in a proper JavaScript project where it is installed via npm (included as a dependency). This could be one of the many examples you have already, merely with a slightly different setup.

Absolutely. I'm also working on serialisation options for the future, which is especially useful for projects that are ran from the command line :)
