### 2022-01-24 53bcb1d
Updated some developer dependencies (jsdoc, rollup, mocha) in packages.json
