### 2022-01-24 870d96a
Updates bundling script to use local packages
