### 2022-01-24 889ca61
Updated readme and bundling script so it better reflects state of published bundle
