+++ make_bundle.sh
@@ -66,8 +66,8 @@ compile_cacatoo()
 
     sed -i 's/images\/elephant_cacatoo_small.png/cacatoo\/images\/elephant_cacatoo_small.png/g' docs/scripts/cacatoo.js
     sed -i 's/images\/gh.png/cacatoo\/images\/gh.png/g' docs/scripts/cacatoo.js           
-    echo "[OK]"    
-    if [[ $1 != "once" ]]; then
+    echo "[OK]"  
+    if [ "$1" != "once" ]; then
         echo -en "Awaiting changes in the code..."
     fi
 }
@@ -75,7 +75,7 @@ compile_cacatoo()
 
 chsum1=""
 
-compile_cacatoo
+compile_cacatoo $1
 
 while [[ true ]]
 do    
