### 2022-01-24 65b6e7a
Updated readme and bundling script so it better reflects state of published bundle
