### 2022-01-24 71d4e2f
Started implementing checkpoints. Testing in cheater
