+++ src/simulation.js
@@ -41,6 +41,25 @@ class Simulation {
         if(config.printcursor == false) this.printcursor = false        
     }
 
+    save_checkpoint() 
+    {
+        console.log("Saving checkpoint of simulation at time ", this.time)
+        return JSON.stringify(this)
+    }
+
+    load_checkpoint(str, type)
+    {
+        console.log("Reloading  checkpoint from string")
+        let revived_sim = new type() 
+        let parser = JSON.parse(str);     
+        Object.assign(revived_sim, parser);
+        // //let revived_persons = []
+        // //for(let i of parsed_house.persons) 
+        // //    revived_persons.push(new Person(i.name,i.age))      
+        // //revived_house.persons = revived_persons
+        return revived_sim;  
+    }
+
     /**
     *  Generate a new GridModel within this simulation.  
     *  @param {string} name The name of your new model, e.g. "gol" for game of life. Cannot contain whitespaces. 
