+++ cacatodo.txt
@@ -1,3 +1,4 @@
 (caca)TODO
+> unit testing for checkpointing
 > FBA testing
 > Implement d3.js visualization like histograms / phylogenetic trees. 
