### 2022-01-26 2fd0990
First step towards serialisation
