### 2022-01-26 #3948 @Bisaloo
Great! Thank you again @emilydolson and @TimKam for your valuable comments!

@bramvandijk88, I'll now do a last review for typos. I encourage you to do the same because there is no proofing after acceptance at JOSS. Once this is done, we can take the final steps to accept this submission.
