### 2022-01-26 #3948 @Bisaloo
Shouldn't https://github.com/bramvandijk88/cacatoo/blob/65b6e7a2daac8175d65a6052b4f2f6d3652d08b7/paper/paper.bib#L1-L10 be `@article` instead of `@misc` anyways?
