+++ paper/paper.md
@@ -20,23 +20,22 @@ bibliography: paper.bib
 
 # Summary
 
-In the field of ecology and evolution, a treasure trove of data has revealed the importance of spatial structure and biogeography. Despite these rich data sets, our conceptual understanding of how spatial structure shapes biodiversity, pathogenicity, and microbial pangenomes is lagging behind. For example, we only have a limited understanding of how interactions at the microscale (molecular machinery, bacteriophages, metabolism) scale up to define eco-evolutionary dynamics of microbial communities [@rainey2020toward] and metaorganisms [@jaspers2019resolving]. To develop our intuition on these systems, I argue we need to embrace multiple levels and structural complexity and in our models. `Cacatoo` is a toolbox developed to make it easy to design, explore, and share simulations of such systems. Simulations can be interactively explored from a web browser, allowing the user to change parameters and observe graphs in real time. `Cacatoo` is designed to be both easy-to-use and extendible, making it suitable for beginners and experts alike. Because it requires no installation and works on practically every computer, it is also ideal for teaching purposes and student projects. In summary, `Cacatoo` provides opportunities for everyone to get involved in spatially structured modelling.
+In the field of ecology and evolution, a treasure trove of data has revealed the importance of spatial structure and biogeography. Despite these rich data sets, our conceptual understanding of how spatial structure shapes biodiversity, pathogenicity, and microbial pangenomes is lagging behind. For example, we only have a limited understanding of how interactions at the microscale (molecular machinery, bacteriophages, metabolism) scale up to define eco-evolutionary dynamics of microbial communities [@rainey2020toward] and metaorganisms [@jaspers2019resolving]. To develop our intuition on these systems, I argue we need to embrace multiple levels and structural complexity in our models. `Cacatoo` is a toolbox developed to make it easy to design, explore, and share simulations of such systems. Simulations can be interactively explored from a web browser, allowing the user to change parameters and observe graphs in real-time. `Cacatoo` is designed to be both easy-to-use and extendable, making it suitable for beginners and experts alike. Because it requires no installation and works on practically every computer, it is also ideal for teaching purposes and student projects. In summary, `Cacatoo` provides opportunities for everyone to get involved in spatially structured modelling.
 
 # Statement of need
 
[…]
