+++ paper/paper.bib
@@ -5,6 +5,7 @@
   volume={377},
   number={1842},
   pages={20200460},
+  doi={10.1098/rstb.2020.0460},
   year={2022},
   publisher={The Royal Society}
 }
