### 2022-01-26 #3948 @bramvandijk88
I've added the DOI of "Introduction: the secret lives of Mobile genetic Elements".

Google indexed that citation as @misc, probably because it's not an article in the strict sense, it's just an overview of the content in the special issue.
