### 2022-01-29 0352d73
Fixed an issue with updating time when using step-function
