+++ src/simulation.js
@@ -242,6 +242,7 @@ class Simulation {
     step() {
         for (let i = 0; i < this.gridmodels.length; i++)
             this.gridmodels[i].update()
+        sim.time++
     }
 
     /**
@@ -304,7 +305,6 @@ class Simulation {
                         if(!sim.pause==true){
                             sim.step()
                             sim.events()
-                            sim.time++
                         }
                         let endTime = performance.now();
                         t += (endTime - startTime);
@@ -321,7 +321,6 @@ class Simulation {
                     if (!sim.pause == true) {
                         sim.step()
                         sim.events();
-                        sim.time++
                     }
                     sim.display()
                     if(sim.fpsmeter) meter.tick()
