### 2022-01-29 7a83611
Fixed small issue with timestep updating
