### 2022-01-29 a3aefef
Updated unit test to catch issue
