### 2022-01-29 9c0e0b1
Updated unit test to catch issue
