+++ README.md
@@ -80,7 +80,7 @@ If you want to test your code, or want to issue a pull request, please use Mocha
 If you made a nice Cacatoo model which you would like to see on the [JS fiddle examples page](https://bramvandijk88.github.io/cacatoo/examples_jsfiddle.html), be sure to:
 * Make sure your code works as intended in JS fiddle (should not require any rewriting, just some copy-pasting)
 * Give a title and description of your model 
-* Give your name so I can credit you
+* Include your name so I can credit you
 
 ## **Useful developer commands for contributors**
 
