### 2022-02-01 #3948 @Bisaloo
@bramvandijk88, there is a mismatch in the version listed in Zenodo and the actual tag in GitHub. If you can easily fix it, it would be great. But if it's too difficult, then we can ignore it and proceed.

Additionally, could you please edit the license on Zenodo. It's listed as 'Other' when it should be 'GPL-3'?
