### 2022-02-01 #3948 @bramvandijk88
What mismatch is there? They are both 1.0.1 as far as I can tell.

I've changed the licence to "GPL-3" in the zenodo.json file, but it only updates this upon republishing the package. Zenodo complains when I do that without bumping the version, but I actually don't want to bump the version without having any changes in the code. Is there a way around this?
