### 2022-02-01 #3948 @Bisaloo
Here is what I see. And the link to 1.0.3 on GitHub doesn't work because this version doesn't exist.

![image](https://user-images.githubusercontent.com/10783929/151954428-00e93aac-e77f-4c3c-9be2-5bf8df0530bd.png)
