### 2022-02-01 #3948 @bramvandijk88
Not sure what's going on here. V1.0.3 was the result of me testing the metadata, for which I was forced to publish (or Zenodo wouldn't reindex them).

I'm not sure how to resolve this. There's no way to edit these things after publication as far as I can tell, but the DOI-link on the github page itself should get you to version 1.0.1.
