+++ src/simulation.js
@@ -34,10 +34,10 @@ class Simulation {
         this.canvases = []              // Array with refs to all canvases (from all models) from this simulation
         this.graphs = []                // All graphs
         this.time = 0
-        this.inbrowser = (typeof document !== "undefined")
-        this.printcursor = true
+        this.inbrowser = (typeof document !== "undefined")        
         this.fpsmeter = false
         if(config.fpsmeter == true) this.fpsmeter = true
+        this.printcursor = true
         if(config.printcursor == false) this.printcursor = false        
     }
 
@@ -93,12 +93,21 @@ class Simulation {
     
     /**
     * Create a display for a gridmodel, showing a certain property on it. 
-    * @param {string} name The name of an existing gridmodel to display
+    * @param {object} config Object with the keys name, property, label, width, height, scale, minval, maxval, nticks, decimals, num_colours, fill
+    *                        These keys:value pairs are:
+    * @param {string} name The name of the model to display
     * @param {string} property The name of the property to display
     * @param {string} customlab Overwrite the display name with something more descriptive
     * @param {integer} height Number of rows to display (default = ALL)
     * @param {integer} width Number of cols to display (default = ALL)
     * @param {integer} scale Scale of display (default inherited from @Simulation class)
+    * @param {numeric}  minval colour scale is capped off below this value
+    * @param {numeric}  maxval colour scale is capped off above this value
+    * @param {integer} nticks how many ticks
+    * @param {integer} decimals how many decimals for tick labels
+    * @param {integer} num_colours how many steps in the colour gradient
+    * @param {string} fill type of gradient to use (viridis, inferno, red, green, blue)
+
     */
     createDisplay_continuous(config) {  
         if(! this.inbrowser) {
@@ -167,9 +176,6 @@ class Simulation {
             console.warn("Cacatoo:spaceTimePlot, cannot create display in command-line mode.")
             return
         }
-
-        // this.createDisplay(name, property, customlab, height, width, scale)
-        //name, property, customlab, height, width, scale
         
         let source_canvas = this[name].canvases[source_canvas_label]
         let property = source_canvas.property
@@ -223,6 +229,7 @@ class Simulation {
     * @param {scale} scale The zoom (scale) of the grid to grab the correct grid point
     */
     printCursorPosition(canvas, event, scale){
+        if(!this.printcursor) return
         let coords = this.getCursorPosition(canvas,event,scale)
         let x = coords.x
         let y = coords.y
