### 2022-02-04 94d3ca4
Added config-object based display creation to be more consistent with continuous displays
